Ajustar validações para executar ações no arquivo

+++ FileShare.Common/FileShare.Common/SerializableActions/RenameFileAction.cs
@@ -12,7 +12,7 @@ namespace FileShare.Common.SerializableActions
 
         public override void Run()
         {
-            if (File.Exists(OldFileName))
+            if (File.Exists(OldFileName) && FileName != OldFileName)
                 File.Move(OldFileName, FileName);
         }
     }
